Refactor infer typing.Self logic
